files/requirements.txt
# ==============================================================================
#  SGB II MaEnde - Python Dependencies
#
#  Install all dependencies with:
#      pip install -r requirements.txt
#
#  Run tests with:
#      pip install -r requirements.txt
#      python -m pytest tests/ -v
# ==============================================================================

# MicroStrategy Python SDK — core library for all MSTR API calls
# Docs: https://github.com/MicroStrategy/mstrio-py
mstrio-py>=11.3.0

# Loads credentials from deployment.env into the application
python-dotenv>=1.0.0

# ==============================================================================
#  Development / Testing only (not needed in production)
# ==============================================================================

# Test framework — run with: python -m pytest tests/ -v
pytest>=7.0.0
